refactor(user): split avatar internal endpoints into new user/ app
Scope: code review on PR #231 asked the avatar endpoints to live with
their domain instead of growing app/views/internal.py.

* New `user` Django app, scaffolded with manage.py startapp.
  No models yet — the planned profile-adapter epic (docs/plans/
  2026-05-17-user-profile-entity-adapter-epic-design.md) will move
  CustomUser fields here in a follow-up. Migration-free for now.

* Moved verbatim from app/views/internal.py → user/internal/avatar.py:
    - list_discord_linked_users
    - list_discord_guild_ids
    - bulk_update_user_avatars
    - _validated_avatar_updates helper

* backend/urls.py imports them from the new path.
  app/views/internal.py keeps the discord lease-sweep endpoint.

* settings.py registers user.apps.UserConfig in INSTALLED_APPS so
  Django auto-discovers the app for migrate.

* backend/tasks.py db_makemigrations adds "user" to the apps tuple
  so `just db::migrate::all` includes the new app in every env.

Verification:
  python manage.py check                       → no issues
  python manage.py makemigrations user --dry-run → No changes detected
  just db::migrate::all                        → user listed across all 3 envs
  Touched-suite tests                          → 74/74 OK
