+++ backend/backend/settings.py
@@ -133,6 +133,7 @@ INSTALLED_APPS = [
     "bracket.apps.TournamentConfig",
     "discordbot.apps.DiscordbotConfig",
     "events.apps.EventsConfig",
+    "user.apps.UserConfig",
     "cacheops",  # Added for django-cacheops
 ]
